+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# v0.3.0
+
+- Rewrite the `debugger.dd`.
+- New `terminal_color` class for printing with colors.
+
 # v0.2.1
 
 - Add traceback function to debugger lib.